src/ch12/Profile.java
package ch12;

import java.util.HashMap;
import java.util.Map;

public class Profile {
    private Map<String, Answer> answers = new HashMap<>();

    public boolean matches(Criterion criterion) {
        Answer answer = matchingProfileAnswers(criterion);
        return criterion.getAnswer().match(answer);
    }

    public Answer matchingProfileAnswers(Criterion criterion) {
        return answers.get(criterion.getAnswer().getQuestionText());
    }

    public void add(Answer answer) {
        answers.put(answer.getQuestionText(), answer);
    }
}
